requirements.txt
pygit
streamlit
ollama
transformers
torch